Move vue views to separate folder
