+++ resources/js/views/LeagueFixtures.vue
@@ -1,6 +1,7 @@
 <template class="antialiased bg-gray-100 text-gray-600 h-screen px-4" x-data="app">
     <div class="flex flex-col justify-center h-full" v-if="weeks">
         <league-week v-for="week in weeks" :week="week" :key="week.id" />
+        Start Simulation
     </div>
 </template>
