+++ resources/js/views/LeagueTeams.vue
